WozLimit for campaign changed to float and check if its a number and replace comma's with dots

+++ database/migrations/2025_02_18_132700_add_subsidy_possible_and_woz_limit_to_campaigns_table.php
@@ -15,7 +15,7 @@ class AddSubsidyPossibleAndWozLimitToCampaignsTable extends Migration
     {
         Schema::table('campaigns', function (Blueprint $table) {
             $table->boolean('subsidy_possible')->default(false);
-            $table->integer('woz_limit')->nullable();
+            $table->double('woz_limit')->nullable();
         });
 
     }
